CLOSE TDEV-72 add drag and drop for task

+++ src/services/taskService.js
@@ -23,9 +23,11 @@ function addTask(project, type, name, description, usId, time, dependencies) {
             return reject(new Error('userStory parameter is required'));
         if (!time)
             time = 0;
+        if(!dependencies)
+            dependencies = []
 
         const index = TypeValue[type] + '-' + (project.tasks.length+1);
-        let task = new Task({id:index, name:name, description:description, userStoryID:usId, timeEstimation:time, dependencies:dependencies});
+        let task = new Task({id:index, type:type, name:name, description:description, userStoryID:usId, timeEstimation:time, dependencies:dependencies});
         addTaskInUserStory(task, usId)
             .then(() => {
                 project.tasks.push(task);
@@ -93,11 +95,17 @@ function deleteTask(project, _id){
         if (!project){
             return reject(new Error('project parameter is required'));
         }
-        Task.deleteOne({_id:_id, assignee:undefined}).then(value => {
-            if (value.deletedCount === 0)
-                return reject(new Error("UserStory don't delete"));
-            resolve(project.tasks.pull(_id));
-        })
+        Task.find({dependencies:{ "$in" : [_id]} })
+            .then(value => {
+                if(value.length > 0)
+                    return reject(new Error('task is in dependencies'));
+                Task.deleteOne({_id:_id, assignee:undefined}).then(value => {
+                    if (value.deletedCount === 0)
+                        return reject(new Error("UserStory don't delete"));
+                    resolve(project.tasks.pull(_id));
+                })
+                .catch((err) => reject(err));
+            })
             .catch((err) => reject(err));
     });
 }
